Trait math_text_transform::MathTextTransform

pub trait MathTextTransform<T> {
    fn to_math_bold(&self) -> T;
    fn to_math_italic(&self) -> T;
    fn to_math_bold_italic(&self) -> T;
    fn to_math_sans_serif(&self) -> T;
    fn to_math_sans_serif_bold(&self) -> T;
    fn to_math_sans_serif_italic(&self) -> T;
    fn to_math_sans_serif_bold_italic(&self) -> T;
    fn to_math_script(&self) -> T;
    fn to_math_bold_script(&self) -> T;
    fn to_math_fraktur(&self) -> T;
    fn to_math_bold_fraktur(&self) -> T;
    fn to_math_monospace(&self) -> T;
    fn to_math_double_struck(&self) -> T;
}

Convenience trait that allows you to call the the transformation straight on a string slice. If a variant doesn't exist for a given character, it is left as is.

Examples

use math_text_transform::MathTextTransform;

assert_eq!("ItalÄ±c f(Ï€)".to_math_italic(), "ğ¼ğ‘¡ğ‘ğ‘™ğš¤ğ‘ ğ‘“(ğœ‹)");
assert_eq!("Double-struck 123".to_math_double_struck(), "ğ”»ğ• ğ•¦ğ•“ğ•ğ•–-ğ•¤ğ•¥ğ•£ğ•¦ğ•”ğ•œ ğŸ™ğŸšğŸ›");

Required Methods

fn to_math_bold(&self) -> T

Transform to ğ›ğ¨ğ¥ğ variant.

fn to_math_italic(&self) -> T

Transform to ğ‘–ğ‘¡ğ‘ğ‘™ğ‘–ğ‘ variant.

fn to_math_bold_italic(&self) -> T

Transform to ğ’ƒğ’ğ’ğ’… ğ’Šğ’•ğ’‚ğ’ğ’Šğ’„ variant.

fn to_math_sans_serif(&self) -> T

Transform to ğ—Œğ–ºğ—‡ğ—Œ-ğ—Œğ–¾ğ—‹ğ—‚ğ–¿ variant.

fn to_math_sans_serif_bold(&self) -> T

Transform to ğ˜€ğ—®ğ—»ğ˜€-ğ˜€ğ—²ğ—¿ğ—¶ğ—³ ğ—¯ğ—¼ğ—¹ğ—± variant.

fn to_math_sans_serif_italic(&self) -> T

Transform to ğ˜´ğ˜¢ğ˜¯ğ˜´-ğ˜´ğ˜¦ğ˜³ğ˜ªğ˜ ğ˜ªğ˜µğ˜¢ğ˜­ğ˜ªğ˜¤ variant.

fn to_math_sans_serif_bold_italic(&self) -> T

Transform to ğ™¨ğ™–ğ™£ğ™¨-ğ™¨ğ™šğ™ğ™ğ™› ğ™—ğ™¤ğ™¡ğ™™ ğ™ğ™©ğ™–ğ™¡ğ™ğ™˜ variant.

fn to_math_script(&self) -> T

Transform to ğ“ˆğ’¸ğ“‡ğ’¾ğ“…ğ“‰ variant.

fn to_math_bold_script(&self) -> T

Transform to ğ“«ğ“¸ğ“µğ“­ ğ“¼ğ“¬ğ“»ğ“²ğ“¹ğ“½ variant.

fn to_math_fraktur(&self) -> T

Transform to ğ”£ğ”¯ğ”ğ”¨ğ”±ğ”²ğ”¯ variant.

fn to_math_bold_fraktur(&self) -> T

Transform to ğ–‡ğ–”ğ–‘ğ–‰ ğ–‹ğ–—ğ–†ğ–ğ–™ğ–šğ–— variant.

fn to_math_monospace(&self) -> T

Transform to ğš–ğš˜ğš—ğš˜ğšœğš™ğšŠğšŒğš variant.

fn to_math_double_struck(&self) -> T

Transform to ğ••ğ• ğ•¦ğ•“ğ•ğ•–-ğ•¤ğ•¥ğ•£ğ•¦ğ•”ğ•œ variant.